Move fixtures around [no ci]
