+++ emulation_system/tests/compose_file_creator/conversion_logic/service_builders/test_concrete_can_server_service_builder.py
@@ -5,6 +5,7 @@ from typing import Any, Callable, Dict
 import pytest
 from pydantic import parse_obj_as
 from pytest_lazyfixture import lazy_fixture  # type: ignore[import]
+from validation_helper_functions import build_args_are_none, partial_string_in_mount
 
 from emulation_system import OpentronsEmulationConfiguration, SystemConfigurationModel
 from emulation_system.compose_file_creator import BuildItem
@@ -12,10 +13,6 @@ from emulation_system.compose_file_creator.conversion import (
     ConcreteCANServerServiceBuilder,
 )
 from emulation_system.consts import DEV_DOCKERFILE_NAME, DOCKERFILE_NAME
-from tests.compose_file_creator.conversion_logic.conftest import (
-    build_args_are_none,
-    partial_string_in_mount,
-)
 
 
 @pytest.fixture
